+++ core/node/da_dispatcher/src/da_dispatcher.rs
@@ -103,7 +103,7 @@ impl DataAvailabilityDispatcher {
             .await?;
         drop(conn);
 
-        for batch in batches {
+        for batch in &batches {
             let dispatch_latency = METRICS.blob_dispatch_latency.start();
             let dispatch_response = retry(self.config.max_retries(), batch.l1_batch_number, || {
                 self.client
@@ -119,14 +119,14 @@ impl DataAvailabilityDispatcher {
             })?;
             let dispatch_latency_duration = dispatch_latency.observe();
 
-            let sent_at = Utc::now().naive_utc();
+            let sent_at = Utc::now();
 
             let mut conn = self.pool.connection_tagged("da_dispatcher").await?;
             conn.data_availability_dal()
                 .insert_l1_batch_da(
                     batch.l1_batch_number,
                     dispatch_response.blob_id.as_str(),
-                    sent_at,
+                    sent_at.naive_utc(),
                 )
                 .await?;
             drop(conn);
@@ -135,11 +135,38 @@ impl DataAvailabilityDispatcher {
                 .last_dispatched_l1_batch
                 .set(batch.l1_batch_number.0 as usize);
             METRICS.blob_size.observe(batch.pubdata.len());
+            METRICS.sealed_to_dispatched_lag.observe(
+                sent_at
+                    .signed_duration_since(batch.sealed_at)
+                    .to_std()
+                    .context("sent_at has to be higher than sealed_at")?,
+            );
             tracing::info!(
                 "Dispatched a DA for batch_number: {}, pubdata_size: {}, dispatch_latency: {dispatch_latency_duration:?}",
                 batch.l1_batch_number,
                 batch.pubdata.len(),
             );
+        }
+
+        // We don't need to report this metric every iteration, only once when the balance is changed
+        if !batches.is_empty() {
+            let client_arc = Arc::new(self.client.clone_boxed());
+
+            tokio::spawn(async move {
+                let balance = client_arc
+                    .balance()
+                    .await
+                    .with_context(|| "Unable to retrieve DA operator balance");
+
+                match balance {
+                    Ok(balance) => {
+                        METRICS.operator_balance.set(balance);
+                    }
+                    Err(err) => {
+                        tracing::error!("{err}")
+                    }
+                }
+            });
         }
 
         Ok(())